# openapi.yaml
openapi: 3.1.0
info:
  title: Business Management & POS System API
  version: 1.0.0
  summary: REST + JSON contract for the single-business web-based POS and operations platform.
  description: Authoritative machine-readable contract derived from API-Architecture-V1.0.md. Server is the only authority
    for business rules, accounting, and inventory. All derived balances (COGS, payment_state, AR/AP/CRL/SRec, cash_balance,
    inventory_value, moving_average_unit_cost, on_hand_quantity) are server-computed. Posted records are immutable; corrections
    go through lifecycle actions (cancel, return, refund). Cancellation, return, and refund are distinct events. All amounts
    use NUMERIC(15,2); quantities NUMERIC(15,4). Single currency (IDR). All timestamps are ISO 8601 with offset; server normalizes
    to UTC. Session-based auth (Bearer opaque tokens) per §2 of the API Architecture.
  contact:
    name: Business POS System API Team
  license:
    name: Proprietary
servers:
- url: https://api.example.com/api/v1
  description: Production
- url: https://staging-api.example.com/api/v1
  description: Staging
- url: http://localhost:8000/api/v1
  description: Local development
...
paths:
...
  /sales/{id}/returns:
    parameters:
    - name: id
      in: path
      required: true
      description: Sale id
      schema:
        type: integer
        format: int64
    get:
      operationId: listSaleReturns
      summary: List returns on a sale.
      description: List returns on a sale.
      tags:
      - Sales Returns
      security:
      - bearerAuth: []
      x-required-capability: sale.view
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PerPage'
      - $ref: '#/components/parameters/Sort'
      - $ref: '#/components/parameters/Q'
      - $ref: '#/components/parameters/From'
      - $ref: '#/components/parameters/To'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                - pagination
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SalesReturn'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
                  links:
                    $ref: '#/components/schemas/Links'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: createSaleReturn
      summary: 'Process a return: increase stock at original cost; update lifecycle.'
      description: 'Process a return: increase stock at original cost; update lifecycle.'
      tags:
      - Sales Returns
      security:
      - bearerAuth: []
      x-required-capability: sale.return
      parameters:
      - name: id
        in: path
        required: true
        description: Resource id
        schema:
          type: integer
          format: int64
      - $ref: '#/components/parameters/IdempotencyKeyRequired'
      - $ref: '#/components/parameters/IfMatchRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaleReturnRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesReturn'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictErrorEnvelope'
          x-error-codes: lifecycle_state_invalid / return_quantity_exceeds_original / version_mismatch
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /purchases:
    get:
      operationId: listPurchases
      summary: List purchases.
      description: List purchases.
      tags:
      - Purchases
      security:
      - bearerAuth: []
      x-required-capability: purchase.view
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PerPage'
      - $ref: '#/components/parameters/Sort'
      - $ref: '#/components/parameters/Q'
      - $ref: '#/components/parameters/From'
      - $ref: '#/components/parameters/To'
      - name: filter[lifecycle_status]
        in: query
        required: false
        schema:
          type: string
          enum:
          - draft
          - posted
          - completed
          - partially_returned
          - returned
          - cancelled
      - name: filter[supplier_id]
        in: query
        required: false
        schema:
          type: integer
      - name: filter[payment_state]
        in: query
        required: false
        schema:
          type: string
          enum:
          - unpaid
          - partial
          - paid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                - pagination
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Purchase'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
                  links:
                    $ref: '#/components/schemas/Links'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: createPurchase
      summary: Create purchase in Draft. Supplier optional (counter/cash buys allowed).
      description: Create purchase in Draft. Supplier optional (counter/cash buys allowed).
      tags:
      - Purchases
      security:
      - bearerAuth: []
      x-required-capability: purchase.create
      parameters:
      - $ref: '#/components/parameters/IdempotencyKeyRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurchaseCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Purchase'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /purchases/{id}:
    parameters:
    - name: id
      in: path
      required: true
      description: Purchase id
      schema:
        type: integer
        format: int64
    get:
      operationId: getPurchase
      summary: Get purchase. Use `?include=lines,shipping,payments,returns`.
      description: Get purchase. Use `?include=lines,shipping,payments,returns`.
      tags:
      - Purchases
      security:
      - bearerAuth: []
      x-required-capability: purchase.view
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Purchase'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    patch:
      operationId: updatePurchaseDraft
      summary: Update Draft purchase.
      description: Update Draft purchase.
      tags:
      - Purchases
      security:
      - bearerAuth: []
      x-required-capability: purchase.edit_own_draft
      parameters:
      - $ref: '#/components/parameters/IfMatchRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurchasePatch'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          x-error-codes: lifecycle_state_invalid (not draft)
    delete:
      operationId: deletePurchaseDraft
      summary: Hard delete Draft purchase.
      description: Hard delete Draft purchase.
      tags:
      - Purchases
      security:
      - bearerAuth: []
      x-required-capability: purchase.create
      parameters:
      - $ref: '#/components/parameters/IdempotencyKeyRequired'
      responses:
        '204':
          description: No Content
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          x-error-codes: lifecycle_state_invalid (not draft)
        '404':
          $ref: '#/components/responses/NotFound'
...
  /purchases/{id}/cancel:
    post:
      operationId: cancelPurchase
      summary: Cancel a posted purchase. On-hand portion reverses via `purchase_reversal`; consumed portion via `value_adjustment`
        (qty=0).
      description: Cancel a posted purchase. On-hand portion reverses via `purchase_reversal`; consumed portion via `value_adjustment`
        (qty=0).
      tags:
      - Purchases
      security:
      - bearerAuth: []
      x-required-capability: purchase.cancel
      parameters:
      - name: id
        in: path
        required: true
        description: Resource id
        schema:
          type: integer
          format: int64
      - $ref: '#/components/parameters/IdempotencyKeyRequired'
      - $ref: '#/components/parameters/IfMatchRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurchaseCancelRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Purchase'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictErrorEnvelope'
          x-error-codes: already_cancelled / lifecycle_state_invalid / version_mismatch
        '400':
          $ref: '#/components/responses/BadRequest'
  /purchases/{id}/returns:
    parameters:
    - name: id
      in: path
      required: true
      description: Purchase id
      schema:
        type: integer
        format: int64
    post:
      operationId: createPurchaseReturn
      summary: Return goods to a supplier. Stock decreases at original landed cost.
      description: Return goods to a supplier. Stock decreases at original landed cost.
      tags:
      - Purchase Returns
      security:
      - bearerAuth: []
      x-required-capability: purchase.return
      parameters:
      - name: id
        in: path
        required: true
        description: Resource id
        schema:
          type: integer
          format: int64
      - $ref: '#/components/parameters/IdempotencyKeyRequired'
      - $ref: '#/components/parameters/IfMatchRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurchaseReturnRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseReturn'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictErrorEnvelope'
          x-error-codes: lifecycle_state_invalid / purchase_return_quantity_exceeds_original / version_mismatch
        '400':
          $ref: '#/components/responses/BadRequest'
  /purchase-returns:
    get:
      operationId: listPurchaseReturns
      summary: List all purchase returns.
      description: List all purchase returns.
      tags:
      - Purchase Returns
      security:
      - bearerAuth: []
      x-required-capability: purchase.view
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PerPage'
      - $ref: '#/components/parameters/Sort'
      - $ref: '#/components/parameters/Q'
      - $ref: '#/components/parameters/From'
      - $ref: '#/components/parameters/To'
      - name: filter[purchase_id]
        in: query
        required: false
        schema:
          type: integer
      - name: filter[supplier_id]
        in: query
        required: false
        schema:
          type: integer
      - name: filter[lifecycle_status]
        in: query
        required: false
        schema:
          type: string
          enum:
          - posted
          - cancelled
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                - pagination
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PurchaseReturn'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
                  links:
                    $ref: '#/components/schemas/Links'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /purchase-returns/{id}:
    parameters:
    - name: id
      in: path
      required: true
      description: Purchase return id
      schema:
        type: integer
        format: int64
    get:
      operationId: getPurchaseReturn
      summary: Get purchase return.
      description: Get purchase return.
      tags:
      - Purchase Returns
      security:
      - bearerAuth: []
      x-required-capability: purchase.view
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseReturn'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /purchase-returns/{id}/cancel:
    post:
      operationId: cancelPurchaseReturn
      summary: 'Cancel a purchase return: insert reversal movements, update lifecycle.'
      description: 'Cancel a purchase return: insert reversal movements, update lifecycle.'
      tags:
      - Purchase Returns
      security:
      - bearerAuth: []
      x-required-capability: purchase.return
      parameters:
      - name: id
        in: path
        required: true
        description: Resource id
        schema:
          type: integer
          format: int64
      - $ref: '#/components/parameters/IdempotencyKeyRequired'
      - $ref: '#/components/parameters/IfMatchRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurchaseReturnCancelRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseReturn'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /reports/purchases:
    get:
      operationId: purchasesReport
      summary: Purchase report.
      description: Purchase report.
      tags:
      - Reports
      security:
      - bearerAuth: []
      x-required-capability: report.view
      parameters:
      - $ref: '#/components/parameters/Period'
      - $ref: '#/components/parameters/From'
      - $ref: '#/components/parameters/To'
      - $ref: '#/components/parameters/CompareTo'
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PerPage'
      - $ref: '#/components/parameters/Sort'
      - $ref: '#/components/parameters/Q'
      - name: filter_supplier_id
        in: query
        required: false
        schema:
          oneOf:
          - type: string
          - type: integer
        description: 'Filter parameter (alias: `filter[supplier_id]`).'
      - name: filter_lifecycle_status
        in: query
        required: false
        schema:
          oneOf:
          - type: string
          - type: integer
        description: 'Filter parameter (alias: `filter[lifecycle_status]`).'
      - name: filter_posted_by
        in: query
        required: false
        schema:
          oneOf:
          - type: string
          - type: integer
        description: 'Filter parameter (alias: `filter[posted_by]`).'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchasesReportResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /reports/purchase-returns:
    get:
      operationId: purchaseReturnsReport
      summary: Purchase return report.
      description: Purchase return report.
      tags:
      - Reports
      security:
      - bearerAuth: []
      x-required-capability: report.view
      parameters:
      - $ref: '#/components/parameters/Period'
      - $ref: '#/components/parameters/From'
      - $ref: '#/components/parameters/To'
      - $ref: '#/components/parameters/CompareTo'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Opaque
      description: 'Session-token auth per API Architecture §2. Server-issued opaque random tokens after credential exchange.
        256-bit random, base64url-encoded, stored hashed in the sessions table. Client does NOT parse them. Send `Authorization:
        Bearer <token>`. 401 = missing/expired/revoked. 403 = authenticated but missing capability.'
  parameters:
...
    IdempotencyKeyRequired:
      name: Idempotency-Key
      in: header
      required: true
      description: Required. See `IdempotencyKey`.
      schema:
        type: string
        format: uuid
        maxLength: 100
...
    IfMatchRequired:
      name: If-Match
      in: header
      required: true
      description: Required for lifecycle actions. See `IfMatch`.
      schema:
        type: string
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      description: 1-based page number.
    PerPage:
      name: per_page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 50
      description: Page size. Hard cap 500.
    Sort:
      name: sort
      in: query
      required: false
      schema:
        type: string
      description: Comma-separated list. Prefix `-` for descending.
    Q:
      name: q
      in: query
      required: false
      schema:
        type: string
        maxLength: 200
      description: Case-insensitive substring search.
    From:
      name: from
      in: query
      required: false
      schema:
        type: string
        format: date-time
      description: ISO 8601 lower bound.
    To:
      name: to
      in: query
      required: false
      schema:
        type: string
        format: date-time
      description: ISO 8601 upper bound.
...
  responses:
    BadRequest:
      description: Validation failure (400).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorEnvelope'
    Unauthorized:
      description: Missing, invalid, or expired credentials (401).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: Authenticated but missing required capability (403).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuthorizationErrorEnvelope'
    NotFound:
      description: Resource not found (404).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: State conflict (409).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ConflictErrorEnvelope'
...
  schemas:
...
    ConflictErrorEnvelope:
      allOf:
      - $ref: '#/components/schemas/ErrorEnvelope'
      - type: object
        description: Returned for state conflicts (409).
        properties:
          error:
            type: object
            properties:
              code:
                type: string
                enum:
                - insufficient_stock
                - negative_stock_disallowed
                - payment_exceeds_total
                - allocation_exceeds_payable
                - refund_exceeds_crl
                - repayment_exceeds_srec
                - cash_insufficient
                - return_quantity_exceeds_original
                - purchase_return_quantity_exceeds_original
                - lifecycle_state_invalid
                - already_cancelled
                - cannot_cancel_after_return
                - production_inputs_exceed_stock
                - production_finished_cost_invalid
                - cost_type_deactivated
                - product_deactivated
                - concurrent_modification
                - serialization_failure
                - idempotency_violation
                - delete_not_permitted
                - referenced_by_history
                - version_mismatch
              details:
                type: object
                additionalProperties: true
    Pagination:
      type: object
      required:
      - page
      - per_page
      - total
      - total_pages
      properties:
        page:
          type: integer
          minimum: 1
          example: 1
        per_page:
          type: integer
          minimum: 1
          maximum: 500
          example: 50
        total:
          type: integer
          minimum: 0
          example: 1234
        total_pages:
          type: integer
          minimum: 0
          example: 25
    Links:
      type: object
      properties:
        self:
          type: string
          format: uri-reference
        next:
          type:
          - string
          - 'null'
          format: uri-reference
        prev:
          type:
          - string
          - 'null'
          format: uri-reference
...
    SaleReturnRequest:
      type: object
      required:
      - reason
      - lines
      properties:
        reason:
          type: string
          minLength: 1
          maxLength: 1000
        lines:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/SaleReturnLineRequest'
    SalesReturn:
      type: object
      required:
      - id
      - sale_id
      - return_date
      - total_selling_price_returned
      - total_cost_returned
      - lifecycle_status
      - lines
      - created_at
      - created_by
      properties:
        id:
          type: integer
          format: int64
        sale_id:
          type: integer
          format: int64
        return_date:
          type: string
          format: date-time
        reason:
          type:
          - string
          - 'null'
        total_selling_price_returned:
          type: number
          exclusiveMinimum: 0
        total_cost_returned:
          type: number
          exclusiveMinimum: 0
        lifecycle_status:
          type: string
          enum:
          - posted
          - cancelled
        lines:
          type: array
          items:
            $ref: '#/components/schemas/SalesReturnLine'
        created_at:
          type: string
          format: date-time
        created_by:
          type: integer
...
    Purchase:
      type: object
      required:
      - id
      - purchase_date
      - total_amount
      - lifecycle_status
      - paid_amount
      - outstanding
      - payment_state
      - created_at
      - updated_at
      - version
      properties:
        id:
          type: integer
          format: int64
        reference_no:
          type:
          - string
          - 'null'
          maxLength: 50
        supplier_id:
          type:
          - integer
          - 'null'
          minimum: 1
          description: Optional. Counter/cash buys allowed.
        supplier:
          $ref: '#/components/schemas/Contact'
        purchase_date:
          type: string
          format: date-time
        received_date:
          type:
          - string
          - 'null'
          format: date-time
        total_amount:
          type: number
          minimum: 0
          description: Σ line_total + shipping.amount.
        lifecycle_status:
          type: string
          enum:
          - draft
          - posted
          - completed
          - partially_returned
          - returned
          - cancelled
        payment_state:
          type: string
          enum:
          - unpaid
          - partial
          - paid
        paid_amount:
          type: number
        outstanding:
          type: number
        ap:
          type: number
        supplier_receivable:
          type: number
        posted_at:
          type:
          - string
          - 'null'
          format: date-time
        posted_by:
          type:
          - integer
          - 'null'
        cancellation_date:
          type:
          - string
          - 'null'
          format: date-time
        cancellation_reason:
          type:
          - string
          - 'null'
        cancelled_by:
          type:
          - integer
          - 'null'
        notes:
          type:
          - string
          - 'null'
          maxLength: 2000
        lines:
          type: array
          items:
            $ref: '#/components/schemas/PurchaseLine'
        shipping:
          $ref: '#/components/schemas/PurchaseShipping'
        payments:
          type: array
          items:
            $ref: '#/components/schemas/PurchasePayment'
        returns:
          type: array
          items:
            $ref: '#/components/schemas/PurchaseReturn'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        created_by:
          type: integer
        version:
          type: integer
...
    PurchaseCreateRequest:
      type: object
      required:
      - lines
      properties:
        supplier_id:
          type:
          - integer
          - 'null'
          minimum: 1
        purchase_date:
          type: string
          format: date-time
        notes:
          type: string
          maxLength: 2000
        reference_no:
          type:
          - string
          - 'null'
          maxLength: 50
        lines:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/PurchaseLineInput'
        shipping:
          $ref: '#/components/schemas/PurchaseShippingRequest'
...
security:
- bearerAuth: []